Move services from shared to data module in preperation for shared lib between API and Frontend
