+++ libs/gamify/badges/src/lib/badges.controller.ts
@@ -1,6 +1,6 @@
 import { Roles, User, UserModel } from '@gamify/auth';
 import { Role } from '@gamify/core';
-import { ApplicationsService, BadgesService } from '@gamify/shared';
+import { ApplicationsService, BadgesService } from '@gamify/data';
 import { BadRequestException, Body, Controller, Delete, Get, NotFoundException, Param, ParseIntPipe, Post, Put, UnauthorizedException } from '@nestjs/common';
 import { CreateBadgeInput } from './dto/create-badge.input';
 import { UpdateBadgeInput } from './dto/update-badge.input';
