+++ libs/gamify/data/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/data.service';
 export * from './lib/data.module';
+export * from './lib/services';
