+++ libs/gamify/data/src/lib/services/applications.service.ts
