+++ libs/gamify/data/src/lib/services/index.ts
@@ -1,0 +1,3 @@
+export * from './applications.service';
+export * from './badges.service';
+export * from './users.service';
