--- libs/gamify/shared/README.md
@@ -1,7 +1,0 @@
-# gamify-shared
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test gamify-shared` to execute the unit tests via [Jest](https://jestjs.io).
