--- libs/gamify/shared/project.json
@@ -1,22 +1,0 @@
-{
-  "root": "libs/gamify/shared",
-  "sourceRoot": "libs/gamify/shared/src",
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/gamify/shared/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/gamify/shared"],
-      "options": {
-        "jestConfig": "libs/gamify/shared/jest.config.js",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": ["scope:gamify", "type:util"]
-}
