--- libs/gamify/shared/src/index.ts
@@ -1,4 +1,0 @@
-export * from './lib/gamify-shared.module';
-export * from './lib/services/applications.service';
-export * from './lib/services/badges.service';
-export * from './lib/services/users.service';
