--- libs/gamify/shared/src/lib/gamify-shared.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-
-@Module({
-  controllers: [],
-  providers: [],
-  exports: [],
-})
-export class GamifySharedModule {}
